requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Time-series & signal processing
statsmodels>=0.14.0
pyts>=0.13.0          # time-series transforms

# Imbalanced learning
imbalanced-learn>=0.11.0

# Survival analysis
lifelines>=0.27.0

# Gradient boosting
xgboost>=2.0.0
lightgbm>=4.0.0

# Explainability
shap>=0.43.0

# HuggingFace ecosystem
huggingface_hub>=0.20.0
transformers>=4.35.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# App / demo
gradio>=4.0.0

# Config management
pyyaml>=6.0

# Data download
kaggle>=1.5.16
requests>=2.31.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
einops>=0.7.0          # Transformer tensor ops

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.1.0
